src/requirements.txt

asgiref==3.7.2
celery==5.3.1
Django==4.2.4
django-filter==23.2
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
kombu==5.3.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
Pillow
python-certifi-win32
certifi